backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
httpx==0.27.0
python-dotenv==1.0.1

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── Database ───────────────────────────────────────────────────
sqlalchemy==2.0.30
# psycopg2-binary==2.9.9  # uncomment for PostgreSQL

# ── AI (Modules 3, 4, 8, 14, 15) ──────────────────────────────
openai==1.30.1

# ── NLP (Module 5) ─────────────────────────────────────────────
spacy==3.7.4

# ── Sentiment / Embeddings (Modules 6, 9, 10) ─────────────────
transformers==4.41.1
torch==2.3.0
sentence-transformers==3.0.0

# ── ML (Modules 11, 12) ────────────────────────────────────────
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
